+++ bee-storage/bee-storage-test/src/message_id_to_metadata.rs
@@ -57,7 +57,7 @@ pub async fn message_id_to_metadata_access<B: StorageBackend>(storage: &B) {
         .await
         .unwrap();
     assert_eq!(results.len(), 1);
-    assert!(results[0].is_none());
+    assert!(matches!(results.get(0), Some(Ok(None))));
 
     Insert::<MessageId, MessageMetadata>::insert(storage, &message_id, &metadata)
         .await
@@ -79,7 +79,7 @@ pub async fn message_id_to_metadata_access<B: StorageBackend>(storage: &B) {
         .await
         .unwrap();
     assert_eq!(results.len(), 1);
-    assert_eq!(results[0].as_ref().unwrap(), &metadata);
+    assert!(matches!(results.get(0), Some(Ok(Some(v))) if v == &metadata));
 
     Delete::<MessageId, MessageMetadata>::delete(storage, &message_id)
         .await
@@ -100,7 +100,7 @@ pub async fn message_id_to_metadata_access<B: StorageBackend>(storage: &B) {
         .await
         .unwrap();
     assert_eq!(results.len(), 1);
-    assert!(results[0].is_none());
+    assert!(matches!(results.get(0), Some(Ok(None))));
 
     let mut batch = B::batch_begin();
     let mut message_ids = Vec::new();
@@ -142,8 +142,8 @@ pub async fn message_id_to_metadata_access<B: StorageBackend>(storage: &B) {
 
     assert_eq!(results.len(), message_ids.len());
 
-    for ((_, metadata), result) in metadatas.iter().zip(results.iter()) {
-        assert_eq!(metadata, result);
+    for ((_, metadata), result) in metadatas.into_iter().zip(results.into_iter()) {
+        assert_eq!(metadata, result.unwrap());
     }
 
     Truncate::<MessageId, MessageMetadata>::truncate(storage).await.unwrap();